backend/requirements.txt
fastapi
uvicorn[standard]
torch
torchvision
timm
pillow
python-multipart
pydantic
grad-cam
pymupdf
Pillow